Remove request in test_is_connected that tries to 'wake' up server select loop
This was required with the pre-tornado implementation since the loop had a
fairly long timeout which would cause server.join() to block, and no way of
waking it without causing network io. The tornado based implementation actually
exits cleanly as soon as stop() is called, causing the unnecessary request to
fail. Due to a race condition, it did not happen all the time, but could have
been quite reliably reproduced by adding a time.sleep() before the ?watchdog
request is made.