Threadsafe wrappers for KATCPResource
The purpose of these wrappers is to allow the straight forward use of
KATCPResource objects by simple synchronous scripts that run on the
main thread. Performance is not a priority, since observation scripts
typically have a very low rate of interaction with the telescope.

Based on ProxyTypes (peak.utils.proxies). The original KATCPResource
objects (request, sensor, container and client types) are wrapped with
proxies that run all callables on the ioloop, blocking the calling
thread using concurrent.future.Future objects.