Add sensor object cache to InspectingClient CM-593
It seem the sensor-management code is somehow leaking memory. However, once a
sensor has been instantiated there is no real need to use that machinery on
every update. As a temporary workaround a sensor object cache is added. If the
sensor is not found the original process is followed. This should result in
memory only being leaked when a sensor has changed or is newly added, not on
every sensor update.

Not sure while it is leaking (will still investigate), but this should 99.99%
 solve the problem :)

When reviewing please check very carefully if there won't be a problem with the
object cache in the case were sensors are removed or updated.