Address memory leak when using until_any() with AsyncState.until_state()
Should also help with CM-593.

The implemented solution depends on the internals of tornado.concurrent.Future,
so won't work with thread-safe futures. Might also break in the future :-/

Tried to implement something 'proper' using weakrefs, but I was not
understanding something there, so could not manage to make it work.