New add_group method on container
This allows the dynamic addition of client groups on the container.
It does this by inserting the new group into the resource spec and updating
the groups property, which allows the higher-level thread-safe container
to wrap the new group properly.

There is no real need to remove groups as they are innocuous if they hang
around.