Avoid blocking python interpreter shutdown or leaking file descriptors CB-1146
When the python interpreter exits it garbage collects all generators that still
exists. Sometimes (usually in tests when the test classes keep references to
everything) tornado coroutines (which are generators) might still exist in a
limbo state after the ioloop on which they were running has stopped. The client
has a finally statement in its main coroutine that tries to shutdown
cleanly. This does not work well when the ioloop has been stopped, and ends up
blocking the python interepreter exit for a second or so. Multiply by the number
of clients, this can be quite long at the end of a test run

By simply using add_callback to run the cleanup, nothing bad happens if the
ioloop is no longer running.