Remove explicit sensor status check in wait()
Since the sensor status is already available within the reading, use
the standard callable condition if you want to check the status as well.
This simplifies the wait() interface and was foreshadowed in the wait()
example in tutorial.rst anyway... :-)

Although the original katcorelib versions of wait() checked for
status == 'nominal' by default, the current default is None (i.e. no
status checking). Removing it therefore does not seem to cause much harm.