Add concurrent_reply decorator (CB-1747)
This adds a decorator @concurrent_reply that can be used on a request
handler that returns a future, to indicate that the request should be
fully concurrent (within the connection). It does this by wrapping the
return value in a ConcurrentReply wrapper object, which is detected and
unwrapped by DeviceServer.handle_request.

It still needs more documentation and unit tests. This is an initial
proof-of-concept.