New until_some helper function
This returns a future that resolves when *some* of the passed futures
resolve, as opposed to *all* (the usual yield on multiple futures)
or *any* (the until_any function). This is in fact the general case
and until_all / until_any can be implemented in terms of until_some.

The function is based around tornado.gen.WaitIterator and takes
similar inputs while adding a timeout and done_at_least (the number of
futures required to resolve). It supports either a list or a dict of
futures, and returns the results of resolved futures.