Let ClientGroup wait for a quorum of clients
Give ClientGroup.wait() the ability to wait until a subset of clients
(a 'quorum') satisfy a sensor condition. This provides a way to e.g.
ignore a broken or stowed receptor while slewing the working receptors
onto a target. The wait times out if a quorum could not be achieved in
the given timeout period. The desired quorum can be specified as either
a number or a fraction of clients.

An additional grace period, which forms part of the overall timeout,
allows for some fine-tuning of the behaviour. All remaining clients
that have not yet satisfied the condition (the 'stragglers') are given
the opportunity to do so within the grace period. This can effectively
result in a full quorum by the time the wait resolves, if all clients
behave as expected.

The default behaviour is unchanged: the group waits for all clients
with a timeout of 5 seconds.

This addresses JIRA ticket MKAIV-534.