Allow sampling strategy to be removed from cache
When using a KATCPClientResource, once a sampling strategy has
been set it gets cached and there is no way to remove it.  It
can be set to 'none', but after disconnect/sync events there
will still be attempts to apply it.  This can cause unnecessary
errors and exceptions to be logged when many sensors that were
being sampled are removed from the server's KATCP interface.  The
errors occur during reapply_sampling_strategies.

The workaround is to provide a mechanism to drop cached sensor
stategies on demand.  This is now available as `drop_sampling_strategy`.

JIRA: CB-2544