Fix possible memory leak when stopping client
If the `DeviceClient`'s ioloop is stopped before the main coroutine
has completed, a reference to the client object remains, and it
cannot be cleaned up by the garbage collector.  This can result
in a memory leak.  This is noticeable in long running applications
that create and destroy client resource containers using a single
ioloop instance.

As background, existing client fuctionality allows it to either create
its own ioloop instance on a new thread (managed), or allow the caller
to provide the ioloop (unmanaged).  This choice typically determines
the thread that makes the call to the `stop()` method.  The unmanaged
case will normally use the same thread as the ioloop, while the managed
case will normally be stopped from a different thread.

The client's stop method now works properly for both cases.  When
using the same thread as the ioloop (unmanaged), the caller should
yield on the new `until_stopped()` method rather than calling `join()`.
When stopping from a different thread (managed), then the opposite
applies - use `join()`, not `until_stopped()`.

It is common to use the inspecting client and resource clients
in the unmanaged ioloop mode, so it is important that they wait
using the new `until_stopped()` method.

Added unit tests to verify the objects can be freed.  They use the
simple approach of checking that a weak reference is cleared.

JIRA: MT-834