Change DeviceClient.ioloop to property
It's value may not be changed directly, so make it a read-only
property to prevent accidents.

The `set_ioloop` method can only be called once before starting.
That, combined with preventing a direct change to the ioloop
attribute, allows the check for a change in ioloop during `_install`
to be removed.

JIRA: MT-834