Send final state update from inspecting client
A user may be waiting on the disconnected state of a resource,
so we send a state change when stopping the inspecting client.
This is noticeable when using the `monitor_resource_sync_state`
function.

A couple of related tests were added.  There is still a possible
memory leak if stopping a resource container while it is syncing,
but a fix hasn't been found yet.  A test was added, but it is marked
to be skipped until such time as the problem is fixed.

JIRA: MT-834