Increase MAX_QUEUE_SIZE to handle more clients
Increase in queue size to cater for more concurrent client connections.
One of our processes has around 90 clients.

Ideally the limit should be a function of the number of connected
clients, but that is more effort.  Note that the limit used to be
just 3, asssuming a single client - this was changed in 6aabfb8bd1e
to allow for "multiple" clients.

JIRA: MT-92