Use correct ioloop for client AsyncEvent objects
A number of `AsyncEvent` objects are created in the client's
`__init__` method, resulting in them getting a reference to the
current ioloop at the time of instantiation.  If a different ioloop
is used when starting the client (which will happen in the "managed"
case), this can cause problems.  E.g., if the original ioloop is
closed before stopping the client, an exception occurs during
`yield self._stopped.until_set`.

The fix is to recreate the `AsyncEvent` objects using the correct
ioloop when starting the client.  Some of the other state variables
that should have been cleared after stopping, for example the protocol
received, are now also re-initialised each time.

Problems with `AsyncEvent` objects initialised in other higher level
classes were not fixed, but should be considered for future work.

JIRA: MT-834