Make Str type more lenient when encoding
Some of CAM's existing code relies on the fact that non-string
types can be sent via Str, or set as String sensor values.  These
issues were found running the katcore and katsim repos' unit tests.
This led to the addition of the `request` and `return_reply` decorated
request handler tests.

The `ensure_native_str` function is now created slightly differntly,
depending on the Python version, which reduces the number of checks
we need to do on each function call.  The `encoding` parameter was
removed, as we always use UTF-8.

JIRA: MT-737