Allow Sensor with units set to None
Some CAM code relies on this, so make sure it still works.

Also changed default value for `units` from empty string to
`None` as it is more consistent with the other kwargs in
`__init__`.  The functionality is unchanged - if units are not
specified, the empty string is used.

JIRA: MT-737