Increase time allowed for test_slow_client
The code works fine, but when Jenkins runs the py27 and py36 tox
tests in parallel we see occassional failures.  The py27 tests run
slightly faster than the py36 tests.  We see the failure if the py27
tests happen to finish while the py36 test is busy with this slow
client test.  In that scenario this part of the test to take 1.3 to
1.4 seconds (compared to 0.3 to 0.7 seconds normally).

We could run the two sets of tests in series, but we save around 45
seconds doing it in parallel.

JIRA: MT-737