Fix parsing problem with long int params
If a float sensor had `long` integer params (upper or lower limit),
the parsing was failing in the inspecting client on Python 2.  A string
like `'9223372036854775808L'` could not be parsed as a float.

This problem was found in CAM's katnodeman application.  In order to
maintain the katcp-python behaviour, we allow integer type values for
the float sensor.

JIRA: MT-737